feat: make components smaller & add AsyncBoundary
