+++ popuniv-react/src/components/features/click/alert.tsx
@@ -1,0 +1,28 @@
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/components/ui/alert-dialog'
+
+interface ClickAlertProps {
+  open: boolean
+  onClick: VoidFunction
+}
+
+export default function ClickAlert({ open, onClick }: ClickAlertProps) {
+  return (
+    <AlertDialog open={open}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>로그인 후 선택 가능합니다 ٩( ᐛ )و</AlertDialogTitle>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogAction onClick={onClick}>확인</AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  )
+}
